requirements.txt
streamlit
pillow
easyocr
numpy
opencv-python-headless